requirements.txt
Django==1.11
#MySQL-python==1.2.5
gunicorn==19.3.0
PyYAML==3.12
South==1.0.2
argparse==1.2.1
certifi==2017.04.17
chardet==3.0.4
#distribute==0.6.24
django-auth-ldap==1.2.13
django-cors-headers==2.1.0
djangorestframework==3.6.3
elasticsearch==5.2.0
xlrd
xlwt
fpconst==0.7.2
html5lib==0.9999999
idna==2.5
#python-ldap==2.4.41
pytz==2017.2
reportlab==3.4.0
requests==2.18.1
six==1.10.0
suds==0.4
#urllib3==1.21.1
wsgiref==0.1.2


#Django==1.11
#gunicorn==19.3.0
#django-auth-ldap
#django-cors-headers
#south
#djangorestframework
#html5lib==0.9999999
#xhtml2pdf
#reportlab
#six
#elasticsearch==5.1.0
#pyyaml
#fpconst==0.7.2
#suds==0.4
#requests==2.18.1

